no SimTK in base data type
