+++ src/for_3D_build/common/data_type.h
@@ -40,8 +40,6 @@ using Matd = Mat3d;
 using AlignedBox = AlignedBox3d;
 using AngularVecd = Vec3d;
 using Rotation = Rotation3d;
-using SimTKVecd = SimTKVec3;
-
 using BoundingBox = BaseBoundingBox<Vec3d>;
 
 template <class DataType, int array_size>
