+++ src/shared/common/scalar_functions.h
@@ -30,6 +30,7 @@
 
 #include "base_data_type.h"
 
+#include <chrono>
 #include <random>
 
 namespace SPH
